RF: Move resulthook code into core.local
