+++ datalad/interface/utils.py
@@ -54,6 +54,11 @@ from datalad.interface.common_opts import eval_params
 from datalad.interface.common_opts import eval_defaults
 from .results import known_result_xfms
 from datalad.config import ConfigManager
+from datalad.core.local.resulthooks import (
+    get_hooks_from_config,
+    match_hook2result,
+    run_hook,
+)
 
 
 lgr = logging.getLogger('datalad.interface.utils')
@@ -440,7 +445,7 @@ def eval_results(func):
                     # this ensures that they are executed before
                     # a potentially wrapper command gets to act
                     # on them
-                    if hook_match_result(hook, r, spec['match']):
+                    if match_hook2result(hook, r, spec['match']):
                         lgr.debug('Result %s matches hook %s', r, hook)
                         # a hook is also a command that yields results
                         # so yield them outside too
@@ -631,87 +636,3 @@ def _process_results(
                 continue
 
         yield res
-
-
-def get_hooks_from_config(cfg):
-    hooks = {}
-    for h in (k for k in cfg.keys()
-              if k.startswith('datalad.result-hook.')
-              and k.endswith('.match')):
-        proc = cfg.get('{}.proc'.format(h[:-6]), None)
-        if not proc:
-            lgr.warning(
-                'Incomplete result hook configuration %s in %s' % (
-                    h[:-6], cfg))
-            continue
-        sep = proc.index(' ')
-        hooks[h[20:-6]] = dict(
-            cmd=proc[:sep],
-            args=proc[sep + 1:],
-            match=json.loads(cfg.get(h)),
-        )
-    return hooks
-
-
-def hook_match_result(hook, res, match):
-    for k, v in match.items():
-        # do not test 'k not in res', because we could have a match that
-        # wants to make sure that a particular value is not present, and
-        # not having the key would be OK in that case
-
-        # in case the target value is an actual list, an explicit action 'eq'
-        # must be given
-        action, val = (v[0], v[1]) if isinstance(v, list) else ('eq', v)
-        if action == 'eq':
-            if k in res and res[k] == val:
-                continue
-        elif action == 'neq':
-            if k not in res or res[k] != val:
-                continue
-        elif action == 'in':
-            if k in res and res[k] in val:
-                continue
-        elif action == 'nin':
-            if k not in res or res[k] not in val:
-                continue
-        else:
-            lgr.warning(
-                'Unknown result comparison operation %s for hook %s, skipped',
-                action, hook)
-        # indentation level is intended!
-        return False
-    return True
-
-
-def run_hook(hook, spec, r, dataset_arg):
-    import datalad.api as dl
-    cmd_name = spec['cmd']
-    if not hasattr(dl, cmd_name):
-        # TODO maybe a proper error result?
-        lgr.warning(
-            'Hook %s requires unknown command %s, skipped',
-            hook, cmd_name)
-        return
-    cmd = getattr(dl, cmd_name)
-    # apply potential substitutions on the string form of the args
-    # for this particular result
-    # take care of proper JSON encoding for each value
-    enc = json.JSONEncoder().encode
-    # we have to ensure JSON encoding of all values (some might be Path instances),
-    # we are taking off the outer quoting, to enable flexible combination
-    # of individual items in supplied command and argument templates
-    args = spec['args'].format(
-        # we cannot use a dataset instance directly but must take the
-        # detour over the path location in order to have string substitution
-        # be possible
-        dsarg='' if dataset_arg is None else enc(dataset_arg.path).strip('"')
-        if isinstance(dataset_arg, dl.Dataset) else enc(dataset_arg).strip('"'),
-        # skip any present logger that we only carry for internal purposes
-        **{k: enc(str(v)).strip('"') for k, v in r.items() if k != 'logger'})
-    # now load
-    args = json.loads(args)
-    # only debug level, the hook can issue its own results and communicate
-    # through them
-    lgr.debug('Running hook %s: %s%s', hook, cmd_name, args)
-    for r in cmd(**args):
-        yield r
